Extract api to lib
